requirements.txt
Flask==2.1.1
pandas==1.3.3
cassandra-driver==3.25.0

